# Cleanup voxelization